+++ packages/app-bridge/src/types/index.ts
@@ -20,7 +20,7 @@ export * from './File';
 export * from './FileExtension';
 export * from './FileExtensionSets';
 export * from './FileType';
-export * from './Guidelines';
+export * from './Guideline';
 export * from './GuidelineSearchResult';
 export * from './Language';
 export * from './Notify';
